+++ Source/CS/Desktop/DesktopApp/System/frmVegetableProcessList.cs
@@ -23,7 +23,7 @@ namespace DesktopApp
     public partial class frmVegetableProcessList : DockContent 
     {
         public frmMain frmMain { get; set; }
-        decimal Period; //保质期
+        Double Period; //保质期
         private SysUser User;
         NfcTag nfcTag;
         Bmp2Bmp2Data b2d;
@@ -161,7 +161,7 @@ namespace DesktopApp
                     MesWeightRecord.W_GoodsCode = txtGoodsCode.Text;
                     MesWeightRecord.W_GoodsName = txtGoodsName.Text;
                     MesWeightRecord.W_Batch = txtBatch.Text;
-                    MesWeightRecord.W_Qty = decimal.Parse(txtQty.Text);
+                    MesWeightRecord.W_Qty = Double.Parse(txtQty.Text);
                     MesWeightRecord.W_Unit = txtUnit.Text;
 
                     if (MesWeightRecordBLL.SaveEntity("", MesWeightRecord) > 0)
