+++ Source/CS/Desktop/Model/Dto/MesCollarDetailDto.cs
@@ -41,7 +41,7 @@ namespace Model.Dto
         /// <summary>
         /// 数量
         /// </summary>
-        public decimal C_Qty{ set; get; }
+        public Double C_Qty{ set; get; }
 
         /// <summary>
         /// 批次
