+++ Source/CS/Desktop/Model/Dto/MesInventoryDto.cs
@@ -41,7 +41,7 @@ namespace Model.Dto
         /// <summary>
         /// 数量
         /// </summary>
-        public decimal I_Qty{ set; get; }
+        public Double I_Qty{ set; get; }
 
         /// <summary>
         /// 批次批次以入库时间（yyyymmdd)
