+++ Source/CS/Desktop/Model/Dto/MesRequistDetailDto.cs
@@ -41,7 +41,7 @@ namespace Model.Dto
         /// <summary>
         /// 数量
         /// </summary>
-        public decimal R_Qty{ set; get; }
+        public Double R_Qty{ set; get; }
 
         /// <summary>
         /// 批次
