+++ Source/CS/Desktop/Model/System/MesWeightRecordEntity.cs
@@ -39,11 +39,11 @@ namespace Model
         /// <summary>
         /// 数量
         /// </summary>
-        public decimal W_Qty{ set; get; }
+        public Double W_Qty{ set; get; }
 
         /// 价格
         /// </summary>
-        public decimal W_Price { set; get; }
+        public Double W_Price { set; get; }
 
         /// <summary>
         /// 批次
