+++ Source/CS/Desktop/Model/System/Mes_InPrice.cs
@@ -13,8 +13,8 @@ namespace Model
       public string P_SupplyName { set; get; }
       public string P_GoodsCode { set; get; }
       public string P_GoodsName { set; get; }
-      public decimal P_InPrice { set; get; }
-      public decimal P_Itax { set; get; }
+      public Double P_InPrice { set; get; }
+      public Double P_Itax { set; get; }
       
     }
 }
